feat: store 1RM when adding movement with weight

+++ app/src/main/java/net/martinlundberg/a1repmaxtracker/data/database/model/OneRMEntity.kt
@@ -8,7 +8,7 @@ import java.time.OffsetDateTime
 @Entity
 data class OneRMEntity(
     @PrimaryKey(autoGenerate = true) val oneRMid: Int = 0,
-    val movementId: Int,
+    val movementId: Long,
     val weight: Int,
     val date: OffsetDateTime,
 )
