+++ app/src/main/java/net/martinlundberg/a1repmaxtracker/data/model/OneRMInfo.kt
@@ -14,7 +14,7 @@ data class OneRMInfo(
     }
 }
 
-fun OneRMInfo.asEntity(movementId: Int) = OneRMEntity(
+fun OneRMInfo.asEntity(movementId: Long) = OneRMEntity(
     movementId = movementId,
     weight = weight,
     date = date,
